backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
aiosqlite==0.20.0
greenlet>=3.0.0
python-dotenv==1.0.1
httpx==0.27.0
python-multipart==0.0.9
pydantic==2.9.0
pydantic-settings==2.5.0
aiofiles==24.1.0
Pillow==10.4.0
websockets==12.0
pytest==8.3.0
pytest-asyncio==0.24.0
